requirements.txt
streamlit>=1.32.0
yfinance>=0.2.37
finnhub-python>=2.4.19
pandas>=2.1.0
ta>=0.11.0
anthropic>=0.25.0
supabase>=2.4.0
python-dotenv>=1.0.0
plotly>=5.20.0
requests>=2.31.0
pytz>=2024.1
